requirements.txt
numpy
scipy
scikit-image
opencv-python
pillow
diffusers==0.29.2
transformers==4.40.2
accelerate
huggingface-hub
peft==0.10.0
onnxruntime-gpu
insightface
basicsr
